deploy_temp/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
mysql-connector-python==8.2.0
Werkzeug==3.0.1
reportlab==4.0.8
email-validator==2.1.0.post1
Flask-Mail==0.9.1
